ice: Don't allow more channels than LAN MSI-X available (jsc#SLE-7926).
suse-commit: 089780a0778fd1ecc813a74b67fbd019958e9a3a

+++ drivers/net/ethernet/intel/ice/ice.h
@@ -380,7 +380,7 @@ struct ice_pf {
 	u32 num_avail_sw_msix;	/* remaining MSIX SW vectors left unclaimed */
 	u16 max_pf_txqs;	/* Total Tx queues PF wide */
 	u16 max_pf_rxqs;	/* Total Rx queues PF wide */
-	u32 num_lan_msix;	/* Total MSIX vectors for base driver */
+	u16 num_lan_msix;	/* Total MSIX vectors for base driver */
 	u16 num_lan_tx;		/* num LAN Tx queues setup */
 	u16 num_lan_rx;		/* num LAN Rx queues setup */
 	u16 next_vsi;		/* Next free slot in pf->vsi[] - 0-based! */
